+++ src/Resources/contao/languages/de/default.php
@@ -15,3 +15,5 @@ $lang['mf_editLanguages'] = '<img src="system/themes/flexible/icons/edit.svg"> S
 $lang['mf_closeEditLanguages'] = '<img src="system/themes/flexible/icons/back.svg"> Sprachbearbeitung schließen';
 $lang['mf_translateField'][0] = 'Übersetzen (%s)';
 $lang['mf_translateField'][1] = 'Wählen Sie diese Option, um das Feld für die angegebene Sprache zu übersetzen.';
+$lang['mf_language'][0] = 'Auf Sprache beschränken (Leer = wird immer angezeigt)';
+$lang['mf_language'][1] = 'Wählen Sie die Sprache aus, die gegeben sein muss, damit das Inhaltselement angezeigt wird.';