+++ src/Resources/contao/languages/en/default.php
@@ -1,11 +1,17 @@
 <?php
 
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
 $lang = &$GLOBALS['TL_LANG']['MSC']['multilingualFieldsBundle'];
 
-/**
+/*
  * Misc
  */
-$lang['mf_editLanguages']      = '<img src="system/themes/flexible/icons/edit.svg"> Edit languages';
+$lang['mf_editLanguages'] = '<img src="system/themes/flexible/icons/edit.svg"> Edit languages';
 $lang['mf_closeEditLanguages'] = '<img src="system/themes/flexible/icons/back.svg"> Close language editor';
-$lang['mf_translateField'][0]  = 'Translate "%s" (%s)';
-$lang['mf_translateField'][1]  = 'Click this option in order to translate the field for the given language.';
+$lang['mf_translateField'][0] = 'Translate (%s)';
+$lang['mf_translateField'][1] = 'Click this option in order to translate the field for the given language.';