+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2021-06-18
+
+- fixed palette handling -> now only fields in the current palette are displayed in language edit mode; palettes don't
+  need to be set manually but are retrieved automatically
+- added optional language field for `tl_content` in order to hide certain content elements according to the language
+  (needed for rsce) -> usage is discouraged because of unnecessary duplicate data
+- added a new feature for hiding content elements based on a new field called `mf_language`
+
 ## [0.2.5] - 2021-06-11
 
 - fixed palette handling and visualization